synthetic

+++ benchmark/datasets/femnist/femnist.py
@@ -1,10 +1,9 @@
 import argparse
 import os
 os.sys.path.insert(0, '/Users/densechen/code/benchmark')
-from torchvision import transforms
 
 from benchmark.datasets.simulation_dataset import SimulationDataset
-from benchmark.datasets.utils.transforms import LoadImage, LongTensor
+from benchmark.datasets.utils.transforms import FloatTensor, LongTensor
 
 IMAGE_SIZE = 84
 
@@ -12,7 +11,7 @@ IMAGE_SIZE = 84
 def get_femnit(root, train: bool = True):
     data_root = os.path.join(root, 'train' if train else 'test')
 
-    return SimulationDataset(data_root, LongTensor(), LongTensor())
+    return SimulationDataset(data_root, FloatTensor(), LongTensor())
 
 
 if __name__ == '__main__':
